Fix bugs with Offer and TelegramUser using, and change INT to BIGINT for column userId in bot_message table

+++ src/main/java/com/godeltech/gbf/repository/BotMessageRepository.java
@@ -4,11 +4,9 @@ import com.godeltech.gbf.model.db.BotMessage;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.sql.Timestamp;
-import java.time.LocalDate;
-import java.time.LocalDateTime;
 import java.util.List;
 
-public interface BotMessageRepository extends JpaRepository<BotMessage, Integer> {
+public interface BotMessageRepository extends JpaRepository<BotMessage, Long> {
     List<BotMessage> findAllByUserIdAndChatId(Long telegramId, Long chatId);
     List<BotMessage> findByCreatedAtBefore(Timestamp localDateTime);
 }
